requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coverage==7.0.5
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.4
django-debug-toolbar-force==0.2
django-filter==22.1
django-nine==0.2.7
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
packaging==23.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
urllib3==1.26.14